improve: ui revamps according to design
